ci: silence valgrind errors by adding libc6-dbg package
Here is the full error message from valgrind:

```
Fatal error at startup: a function redirection
which is mandatory for this platform-tool combination
cannot be set up.  Details of the redirection are:

A must-be-redirected function
whose name matches the pattern:      strlen
in an object with soname matching:   ld-linux-x86-64.so.2
was not found whilst processing
symbols from the object with soname: ld-linux-x86-64.so.2

Possible fixes: (1, short term): install glibc's debuginfo
package on this machine.  (2, longer term): ask the packagers
for your Linux distribution to please in future ship a non-
stripped ld.so (or whatever the dynamic linker .so is called)
that exports the above-named function using the standard
calling conventions for this platform.  The package you need
to install for fix (1) is called

  On Debian, Ubuntu:                 libc6-dbg
  On SuSE, openSuSE, Fedora, RHEL:   glibc-debuginfo
```